+++ clid/editmeta.py
@@ -36,13 +36,17 @@ class EditMeta(npy.ActionFormV2):
 
         self.gen = self.add(npy.TitleText, name='Genre', value=self.resolve_genre(self.meta.genre))
         self.nextrely += 1
+        self.dat = self.add(npy.TitleText, name='Date/Year', value=self.meta.date)
+        self.nextrely += 1
         self.tno = self.add(npy.TitleText, name='Track Number',
                             value=str(self.meta.track if self.meta.track != 0 else ''))
+        self.nextrely += 2
+        self.com = self.add(npy.TitleText, name='Comment', value=self.meta.comment)
 
     def set_up_handlers(self):
         super().set_up_handlers()
-        self.handlers['^S'] = self.on_ok
-        self.handlers['^Q'] = self.on_cancel
+        self.handlers['^S'] = self.h_ok
+        self.handlers['^Q'] = self.h_cancel
 
     def resolve_genre(self, num_gen):
         """Convert numerical genre values to readable values. Genre may be
@@ -65,20 +69,42 @@ class EditMeta(npy.ActionFormV2):
         else:
             return num_gen
 
-    def on_cancel(self, char):   # char is for handlers
+    def h_ok(self, char):
+        """Handler to save the tags"""
+        self.on_ok()
+
+    def h_cancel(self, char):
+        """Handler to cancel the operation"""
+        self.on_cancel()
+
+    def on_cancel(self):   # char is for handlers
         """Switch to standard view at once without saving"""
         self.editing = False
         self.parentApp.switchForm("MAIN")
 
-    def on_ok(self, char):   # char is for handlers
+    def on_ok(self):   # char is for handlers
         """Save and switch to standard view"""
+        self.meta.date = self.dat.value
+        # FIXME: error thrown when date is not in the format YYYY-MM-DD
+
         self.meta.title = self.tit.value
         self.meta.album = self.alb.value
+        self.meta.genre = self.gen.value
         self.meta.artist = self.art.value
+        self.meta.comment = self.com.value
         self.meta.album_artist = self.ala.value
 
-        # TODO: make tno accept only numbers(or make a popup if other characters are given)
-        self.meta.track = self.tno.value if self.tno.value != '' else '0'   # automatically converted to int by stagger
+        track = self.tno.value if self.tno.value != '' else '0'   # automatically converted to int by stagger
+        try:
+            int(track)
+        except ValueError:
+            npy.notify_confirm(message='Track Number can only take integer values',
+                               title='Track Number not an integer', editw=0)
+            return None
+        else:
+            self.meta.track = track
+        # FIXME: values of tags are reset to initial when ok is pressed(no prob with ^S)
+
         self.meta.write()
 
         status_meta = '{art} - {alb} - {tno}. {title} '.format(