+++ clid/main.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-__version__ = '0.5.1'
+__version__ = '0.5.2'
 
 import curses
 import npyscreen as npy