+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 Clid's Changelog
 ================
 
+v0.5.2
+------
+
+- [x] Add date tag to editor
+- [x] Add label to Preferences
+- [x] Add comment tag to editor
+- [x] Fix: Genre not being saved
+- [x] Fix: Error thrown when Track No. is str(has to be int)
+- [x] Fix: Error thrown when `OK` or `Cancel` is pressed
+
+
 v0.5.1
 ------
 
+- [x] Add keybindings to save(^S) and cancel(^Q) when editing tags
 - [x] Update status line when page up/down is pressed in the main view
 - [x] Fix: app crashes when up/down, page/up/down keys are pressed if there are no files in display
-- [x] Add keybindings to save(^S) and cancel(^Q) when editing tags